refactor: split backends into modules
