+++ src/awkward/_backends/cupy.py
@@ -1,0 +1,42 @@
+# BSD 3-Clause License; see https://github.com/scikit-hep/awkward-1.0/blob/main/LICENSE
+from __future__ import annotations
+
+from awkward._backends.backend import Backend, KernelKeyType
+from awkward._backends.dispatch import register_backend
+from awkward._kernels import CupyKernel, NumpyKernel
+from awkward._nplikes.cupy import Cupy
+from awkward._nplikes.numpy import Numpy
+from awkward._nplikes.numpylike import NumpyMetadata
+from awkward._typing import Final
+
+np = NumpyMetadata.instance()
+numpy = Numpy.instance()
+
+
+@register_backend(Cupy)
+class CupyBackend(Backend):
+    name: Final[str] = "cuda"
+
+    _cupy: Cupy
+
+    @property
+    def nplike(self) -> Cupy:
+        return self._cupy
+
+    @property
+    def index_nplike(self) -> Cupy:
+        return self._cupy
+
+    def __init__(self):
+        self._cupy = Cupy.instance()
+
+    def __getitem__(self, index: KernelKeyType) -> CupyKernel | NumpyKernel:
+        from awkward._connect import cuda
+
+        cupy = cuda.import_cupy("Awkward Arrays with CUDA")
+        _cuda_kernels = cuda.initialize_cuda_kernels(cupy)
+        func = _cuda_kernels[index]
+        if func is not None:
+            return CupyKernel(func, index)
+        else:
+            raise AssertionError(f"CuPyKernel not found: {index!r}")
