+++ src/awkward/_connect/hist.py
@@ -7,7 +7,7 @@ from __future__ import annotations
 import numpy as np
 
 import awkward as ak
-from awkward._backends.backends import NumpyBackend
+from awkward._backends.numpy import NumpyBackend
 from awkward._typing import Any, Sequence
 
 numpy = NumpyBackend.instance()
