+++ src/awkward/_connect/numba/arrayview_cuda.py
@@ -4,7 +4,7 @@
 from numba import types
 
 import awkward as ak
-from awkward._backends.backends import CupyBackend
+from awkward._backends.cupy import CupyBackend
 
 ########## ArrayView Arguments Handler for CUDA JIT
