+++ src/awkward/record.py
@@ -6,7 +6,10 @@ from collections.abc import Iterable
 
 import awkward as ak
 from awkward._backends.backend import Backend
-from awkward._backends.dispatch import regularize_backend
+from awkward._backends.dispatch import (
+    register_backend_lookup_factory,
+    regularize_backend,
+)
 from awkward._behavior import get_record_class
 from awkward._layout import wrap_layout
 from awkward._nplikes.numpylike import NumpyMetadata
@@ -245,3 +248,13 @@ class Record:
         return Record(
             self._array if array is unset else array, self._at if at is unset else at
         )
+
+
+@register_backend_lookup_factory
+def find_record_backend(obj: type):
+    if issubclass(obj, Record):
+
+        def finder(obj: Record):
+            return obj.backend
+
+        return finder
